+++ Assets/Scripts/Rival1.cs
@@ -19,8 +19,8 @@ public class Rival1 : MonoBehaviour
 	private bool daniado = false;
 	public float damageTime = 0.3f;
 	private float damageTimer = 0f;
-	public int maxSalud = 10;
-	private int saludActual;
+	public float maxSalud = 10f;
+	private float saludActual;
 
 	public string nombre;
 	public Sprite avatar;
@@ -197,7 +197,7 @@ public class Rival1 : MonoBehaviour
 
 	}
 
-	public void recibirDanio(int danio, bool golpeadoPorIzq)
+	public void recibirDanio(float danio, bool golpeadoPorIzq)
 	{
 		if (!isDead)
 		{
@@ -214,7 +214,7 @@ public class Rival1 : MonoBehaviour
 					animator.SetTrigger("matadoPorDer");
 				}
 
-				saludActual = 0;
+				saludActual = 0f;
 				isDead = true;
 				rigidBody.AddRelativeForce(new Vector3(3f, 5f, 0f), ForceMode.Impulse);
 			}
